+++ src/MAUI/Maui.Samples/Samples/Layers/DefineClusteringFeatureReduction/DefineClusteringFeatureReduction.xaml.cs
@@ -15,14 +15,9 @@ using Esri.ArcGISRuntime.Mapping.Popups;
 using Esri.ArcGISRuntime.Portal;
 using Esri.ArcGISRuntime.Reduction;
 using Esri.ArcGISRuntime.Symbology;
-using Esri.ArcGISRuntime.Toolkit.UI.Controls;
-using Esri.ArcGISRuntime.UI.Controls;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
+using Map = Esri.ArcGISRuntime.Mapping.Map;
 
-namespace ArcGIS.WPF.Samples.CustomFeatureClustering
+namespace ArcGIS.Samples.DefineClusteringFeatureReduction
 {
     [ArcGIS.Samples.Shared.Attributes.Sample(
         name: "Custom feature clustering",
@@ -31,12 +26,12 @@ namespace ArcGIS.WPF.Samples.CustomFeatureClustering
         instructions: "Tap the draw clusters button to enable clustering on the feature layer. Interact with the controls to customize clustering feature reduction properties. Tap the map to see the cluster feature count and aggregate fields in the popup.",
         tags: new[] { "aggregate", "bin", "cluster", "group", "merge", "normalize", "popup", "reduce", "renderer", "summarize" })]
     [ArcGIS.Samples.Shared.Attributes.OfflineData("aa44e79a4836413c89908e1afdace2ea")]
-    public partial class CustomFeatureClustering
+    public partial class DefineClusteringFeatureReduction
     {
         private FeatureLayer _layer;
         private ClusteringFeatureReduction _clusteringFeatureReduction;
 
-        public CustomFeatureClustering()
+        public DefineClusteringFeatureReduction()
         {
             InitializeComponent();
             _ = Initialize();
@@ -57,6 +52,16 @@ namespace ArcGIS.WPF.Samples.CustomFeatureClustering
 
             // Set the initial viewpoint to Zurich, Switzerland.
             await MyMapView.SetViewpointAsync(new Viewpoint(47.38, 8.53, 8e4));
+
+            // Hide and nullify an opened popup when user taps screen.
+            PopupBackground.GestureRecognizers.Add(new TapGestureRecognizer
+            {
+                Command = new Command(() =>
+                {
+                    PopupBackground.IsVisible = false;
+                    PopupViewer.Popup = null;
+                })
+            });
         }
 
         private void CreateCustomFeatureReduction()
@@ -82,7 +87,7 @@ namespace ArcGIS.WPF.Samples.CustomFeatureClustering
 
             // Define a default symbol to use for features that do not fall within any of the ranges defined by the class breaks.
             classBreaksRenderer.DefaultSymbol = new SimpleMarkerSymbol() { Color = System.Drawing.Color.Pink };
-
+            
             // Create a new clustering feature reduction using the class breaks renderer.
             _clusteringFeatureReduction = new ClusteringFeatureReduction(classBreaksRenderer);
 
@@ -90,7 +95,7 @@ namespace ArcGIS.WPF.Samples.CustomFeatureClustering
             // The aggregate fields summarize values based on the defined aggregate statistic type.
             _clusteringFeatureReduction.AggregateFields.Add(new AggregateField("Total Residential Buildings", "Residential_Buildings", AggregateStatisticType.Sum));
             _clusteringFeatureReduction.AggregateFields.Add(new AggregateField("Average Building Height", "Most_common_number_of_storeys", AggregateStatisticType.Mode));
-
+            
             // Enable the feature reduction.
             _clusteringFeatureReduction.IsEnabled = true;
 
@@ -116,9 +121,9 @@ namespace ArcGIS.WPF.Samples.CustomFeatureClustering
         }
 
         #region EventHandlers
-        private void DisplayLabelsCheckBox_Checked(object sender, RoutedEventArgs e)
+        private void DisplayLabelsCheckBox_CheckedChanged(object sender, CheckedChangedEventArgs e)
         {
-            if ((bool)(sender as CheckBox).IsChecked)
+            if (DisplayLabelsCheckBox.IsChecked)
             {
                 // Create a label definition with a simple label expression.
                 var simpleLabelExpression = new SimpleLabelExpression("[cluster_count]");
@@ -132,20 +137,17 @@ namespace ArcGIS.WPF.Samples.CustomFeatureClustering
             {
                 _clusteringFeatureReduction.LabelDefinitions.Clear();
             }
-
+            
         }
 
-        private void UpdateClusteringProperties(object sender, RoutedPropertyChangedEventArgs<double> e)
+        private void UpdateClusteringProperties(object sender, EventArgs e)
         {
-            if (_layer != null)
-            {
-                // Update the feature reduction's cluster radius and max scale properties.
-                ((ClusteringFeatureReduction)_layer.FeatureReduction).Radius = ClusterRadiusSlider.Value;
-                ((ClusteringFeatureReduction)_layer.FeatureReduction).MaxScale = MaxScaleSlider.Value;
-            }
+            // Update the feature reduction's cluster radius and max scale properties.
+            ((ClusteringFeatureReduction)_layer.FeatureReduction).Radius = ClusterRadiusSlider.Value;
+            ((ClusteringFeatureReduction)_layer.FeatureReduction).MaxScale = MaxScaleSlider.Value;
         }
 
-        private async void MyMapView_GeoViewTapped(object sender, GeoViewInputEventArgs e)
+        private async void MyMapView_GeoViewTapped(object sender, Esri.ArcGISRuntime.Maui.GeoViewInputEventArgs e)
         {
             // Identify the tapped observation.
             IdentifyLayerResult result = await MyMapView.IdentifyLayerAsync(MyMapView.Map.OperationalLayers.First(), e.Position, 3, true);
@@ -155,29 +157,22 @@ namespace ArcGIS.WPF.Samples.CustomFeatureClustering
 
             // Set the popup and make it visible.
             PopupViewer.Popup = result.Popups.FirstOrDefault();
-            PopupBackground.Visibility = Visibility.Visible;
+            PopupBackground.IsVisible = true;
         }
 
-        private void DrawClustersButton_Clicked(object sender, RoutedEventArgs e)
+        private void DrawClustersButton_Clicked(object sender, EventArgs e)
         {
             // Create a new clustering feature reduction.
             CreateCustomFeatureReduction();
 
             // Show the feature reduction's clustering options.
-            CustomFeatureClusteringOptions.Visibility = Visibility.Visible;
+            DefineClusteringFeatureReductionOptions.IsVisible = true;
 
             // Add an event handler for tap events on the map view.
             MyMapView.GeoViewTapped += MyMapView_GeoViewTapped;
 
             // Hide the draw clusters button.
-            DrawClustersButton.Visibility = Visibility.Collapsed;
-        }
-
-        // Hide and nullify the opened popup when user left clicks.
-        private void PopupBackground_MouseLeftButtonDown(object sender, System.Windows.Input.MouseButtonEventArgs e)
-        {
-            PopupBackground.Visibility = Visibility.Collapsed;
-            PopupViewer.Popup = null;
+            DrawClustersButton.IsVisible = false;
         }
         #endregion
     }
