+++ src/MAUI/Maui.Samples/Samples/Layers/DefineClusteringFeatureReduction/defineclusteringfeaturereduction.jpg
