+++ src/WPF/WPF.Viewer/Samples/Layers/DefineClusteringFeatureReduction/DefineClusteringFeatureReduction.xaml
@@ -1,26 +1,34 @@
-<UserControl x:Class="ArcGIS.WinUI.Samples.CustomFeatureClustering.CustomFeatureClustering"
+﻿<UserControl x:Class="ArcGIS.WPF.Samples.DefineClusteringFeatureReduction.DefineClusteringFeatureReduction"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:esriTK="using:Esri.ArcGISRuntime.Toolkit.UI.Controls"
-             xmlns:esriUI="using:Esri.ArcGISRuntime.UI.Controls">
+             xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013">
     <Grid>
-        <esriUI:MapView x:Name="MyMapView" />
-        <Border Width="auto" Style="{StaticResource BorderStyle}">
+        <esri:MapView x:Name="MyMapView" />
+        <Border Width="auto"
+                Height="auto"
+                Style="{StaticResource BorderStyle}">
             <StackPanel Orientation="Vertical">
                 <Button x:Name="DrawClustersButton"
-                        Click="DrawClustersButton_Click"
+                        Padding="5"
+                        Click="DrawClustersButton_Clicked"
                         Content="Draw clusters" />
-                <StackPanel x:Name="CustomFeatureClusteringOptions"
+                <StackPanel x:Name="DefineClusteringFeatureReductionOptions"
                             Orientation="Vertical"
                             Visibility="Collapsed">
                     <TextBlock Margin="0,0,0,10"
                                FontSize="16"
                                FontWeight="Bold"
                                Text="Clustering properties" />
-                    <Grid Grid.RowDefinitions="auto,auto,auto,auto">
+                    <Grid>
+                        <Grid.RowDefinitions>
+                            <RowDefinition Height="auto" />
+                            <RowDefinition Height="auto" />
+                            <RowDefinition Height="auto" />
+                            <RowDefinition Height="auto" />
+                        </Grid.RowDefinitions>
                         <Grid.ColumnDefinitions>
                             <ColumnDefinition Width="140" />
-                            <ColumnDefinition Width="*" />
+                            <ColumnDefinition Width="100" />
                         </Grid.ColumnDefinitions>
                         <TextBlock Grid.Row="0"
                                    Grid.Column="0"
@@ -30,13 +38,13 @@
                                   Grid.Row="0"
                                   Grid.Column="1"
                                   VerticalAlignment="Center"
-                                  Checked="DisplayLabelsCheckBox_CheckedChanged"
+                                  Checked="DisplayLabelsCheckBox_Checked"
                                   IsChecked="false" />
                         <TextBlock Grid.Row="1"
                                    Grid.Column="0"
                                    VerticalAlignment="Center">
                             <Run Text="Cluster radius: " />
-                            <Run Text="{Binding ElementName=ClusterRadiusSlider, Path=Value, Mode=OneWay}" />
+                            <Run Text="{Binding ElementName=ClusterRadiusSlider, Path=Value, Mode=OneWay, StringFormat=n0}" />
                         </TextBlock>
                         <Slider x:Name="ClusterRadiusSlider"
                                 Grid.Row="1"
@@ -49,7 +57,7 @@
                                    Grid.Column="0"
                                    VerticalAlignment="Center">
                             <Run Text="MaxScale: " />
-                            <Run Text="{Binding ElementName=MaxScaleSlider, Path=Value, Mode=OneWay}" />
+                            <Run Text="{Binding ElementName=MaxScaleSlider, Path=Value, Mode=OneWay, StringFormat=n0}" />
                         </TextBlock>
                         <Slider x:Name="MaxScaleSlider"
                                 Grid.Row="2"
@@ -64,7 +72,7 @@
                                    VerticalAlignment="Center">
                             <Run Text="Current map scale: " />
                             <Run Text="1:" />
-                            <Run Text="{Binding ElementName=MyMapView, Path=MapScale, Mode=OneWay}" />
+                            <Run Text="{Binding ElementName=MyMapView, Path=MapScale, Mode=OneWay, StringFormat=n0}" />
                         </TextBlock>
                     </Grid>
                 </StackPanel>
@@ -72,13 +80,16 @@
         </Border>
         <Grid x:Name="PopupBackground"
               Background="#AA333333"
+              MouseLeftButtonDown="PopupBackground_MouseLeftButtonDown"
               Visibility="Collapsed">
-            <Border HorizontalAlignment="Center" VerticalAlignment="Center">
-                <esriTK:PopupViewer x:Name="PopupViewer"
-                                    MaxWidth="400"
-                                    MaxHeight="400"
-                                    Margin="5"
-                                    Padding="5" />
+            <Border HorizontalAlignment="Center"
+                    VerticalAlignment="Center"
+                    Background="White">
+                <esri:PopupViewer x:Name="PopupViewer"
+                                  MaxWidth="400"
+                                  MaxHeight="400"
+                                  Margin="5"
+                                  Padding="5" />
             </Border>
         </Grid>
     </Grid>
