+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Layers/DefineClusteringFeatureReduction/DefineClusteringFeatureReduction.jpg
